Decouple DbusEventBuffer dependency on ScnIndex
This is an incremental change, trying to keep existing code as much as possible.

Added a state to ScnIndex that marks it as disabled. ScnIndex can be disabled
only of QueuePolicy is BLOCK_ON_WRITE.

Changed some of the methods in ScnIndex to be private, before we proliferate their
use to outsiders. Changed the names of some of the methods for my own understanding.

Removed the use of ScnIndex to get minScn. Instead, DbusEventBuffer tracks _minScn
as a member variable.

Left the usage of ScnIndex.getLargerOffset(). It appears that this never happens
when the policy is BLOCK_ON_WRITE.

Removed the method DbusEventBuffer.getFirstEventTimestamp(), looks like it is not
needed any more.

All fast CI tests pass in local build.